Store Data
The form can now put data in a redux store. When the user logs into
the app, the form checks to see if they have been authenticated. For
now, this is just a basic check to see if the username field is in the
store. If not, it redirects the user to the page so that they can fill
in that field. The username will be put in the store, so that when the
user next goes to the root, they will be able to access the app