Redirect based on login info
The site now redirects based on login info. If the User goes to the root, it checks to see if they are logged in. If they are not logged in at the root, they are redirected to the login page. If they are logged in, they are redirect to their home page. Entering in a specific url will take them to that page